requirements.txt
django
djangorestframework
ipython
django-cors-headers
Pillow